requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Babel==2.0.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
python-dotenv==0.19.0
bleach==4.1.0
markdown==3.3.4
Pillow>=8.3.1
python-dateutil==2.8.2 